# docs/api/openapi-baseline.json
{
  "components": {
    "schemas": {
...
      "APIResponse_AuditLogListResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuditLogListResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "Operation completed successfully",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "title": "APIResponse[AuditLogListResponse]",
        "type": "object"
      },
...
      "APIResponse_ReconciliationReportResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReconciliationReportResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "Operation completed successfully",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "title": "APIResponse[ReconciliationReportResponse]",
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "RFC 7807 Error Response structure.",
        "properties": {
          "code": {
            "default": "PAYDAY_ERROR",
            "description": "Domain-specific error code",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "description": "Human-readable explanation specific to this occurrence",
            "title": "Detail",
            "type": "string"
          },
          "extra": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional debugging or context parameters",
            "title": "Extra"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "URI reference identifying the specific occurrence",
            "title": "Instance"
          },
          "status": {
            "description": "HTTP status code",
            "title": "Status",
            "type": "integer"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "title": {
            "description": "Short, human-readable summary of the problem",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "https://payday.cm/errors/general",
            "description": "URI reference identifying the problem type",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "ReconciliationRequest": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/TransactionChannel"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "partner_records": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PartnerStatementItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partner Records"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "required": [
          "channel"
        ],
        "title": "ReconciliationRequest",
        "type": "object"
      },
...
      "TransactionChannel": {
        "enum": [
          "MTN",
          "ORANGE",
          "UBA"
        ],
        "title": "TransactionChannel",
        "type": "string"
      },
...
      "TransactionStatus": {
        "enum": [
          "PENDING",
          "PROCESSING",
          "SUCCESS",
          "FAILED",
          "REVERSED"
        ],
        "title": "TransactionStatus",
        "type": "string"
      },
      "TransactionType": {
        "enum": [
          "DEPOSIT",
          "WITHDRAW"
        ],
        "title": "TransactionType",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\n# PayDay — Integrated e-Wallet API (Republic of Cameroon)\n\n**EVERYDAY IS A PAYDAY**\n\nPayDay is an integrated, neutral e-wallet bridging **MTN Mobile Money**, **Orange Money**, and **UBA Bank** using the **Triangle Model**.\n\n### Client Integrations:\n- **Flutter Mobile App:** End-customer mobile wallet on Android & iOS.\n- **Angular Landing Page:** Public marketing portal, demo simulator, and fee calculator.\n- **Angular Admin Dashboard:** Operations portal for KYC approvals, ledger audits, wallet freezes, and settlement reconciliation.\n    ",
    "title": "PayDay e-Wallet",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/audit-logs": {
      "get": {
        "description": "Queries immutable system-wide audit logs with actor and entity filters.",
        "operationId": "get_audit_logs_api_v1_admin_audit_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "in": "query",
            "name": "entity_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_AuditLogListResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request — RFC 7807 Problem Details"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized — missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden — insufficient role or ownership"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict — e.g. idempotency or state transition"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests — rate limit exceeded (Retry-After set)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable — e.g. shared state store down"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get System Audit Logs (Admin / Auditor)",
        "tags": [
          "Admin Back-Office"
        ]
      }
    },
    "/api/v1/admin/reconcile": {
      "post": {
        "description": "Executes automated reconciliation comparing internal ledger transactions against external partner settlement records.",
        "operationId": "reconcile_channel_api_v1_admin_reconcile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconciliationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ReconciliationReportResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request — RFC 7807 Problem Details"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized — missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden — insufficient role or ownership"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict — e.g. idempotency or state transition"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests — rate limit exceeded (Retry-After set)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable — e.g. shared state store down"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Channel Settlement Reconciliation (Admin / Auditor)",
        "tags": [
          "Admin Back-Office"
        ]
      }
    },
    "/api/v1/admin/transactions": {
      "get": {
        "description": "Queries all platform transactions across all users with multi-criteria filtering.",
        "operationId": "list_admin_transactions_api_v1_admin_transactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tx_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TransactionType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tx Type"
            }
          },
          {
            "in": "query",
            "name": "channel",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TransactionChannel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TransactionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_AdminTransactionListResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request — RFC 7807 Problem Details"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized — missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden — insufficient role or ownership"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict — e.g. idempotency or state transition"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests — rate limit exceeded (Retry-After set)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable — e.g. shared state store down"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List All Transactions (Admin / Auditor)",
        "tags": [
          "Admin Back-Office"
        ]
      }
    },
...
  }
}